initial client side
