+++ client/src/pages/About/About.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './About.css';
+
+function About() {
+  return <div></div>;
+}
+
+export default About;
