+++ client/src/pages/Homepage/Homepage.css
@@ -1,13 +1,14 @@
-body {
+.week-dates {
   display: flex;
+  flex-wrap: wrap;
   justify-content: center;
   align-items: center;
-  height: 90vh;
 }
-.container {
-  margin: 2rem;
+
+.info {
   display: flex;
-  flex-direction: column;
-  justify-content: center;
+  flex-wrap: wrap;
+  justify-content: space-between;
   align-items: center;
+  width: 100%;
 }
